Fix various 64-bit compiler warnings
This resolves various compiler warnings that were occurring during a 64-bit build.

+++ foo_uie_albumlist/actions.cpp
@@ -4,7 +4,7 @@
 class titleformat_hook_view : public titleformat_hook {
 public:
     bool process_field(
-        titleformat_text_out* p_out, const char* p_name, unsigned p_name_length, bool& p_found_flag) override
+        titleformat_text_out* p_out, const char* p_name, size_t p_name_length, bool& p_found_flag) override
     {
         p_found_flag = false;
         if (m_view) {
@@ -17,7 +17,7 @@ public:
         return false;
     }
 
-    bool process_function(titleformat_text_out* p_out, const char* p_name, unsigned p_name_length,
+    bool process_function(titleformat_text_out* p_out, const char* p_name, size_t p_name_length,
         titleformat_hook_function_params* p_params, bool& p_found_flag) override
     {
         p_found_flag = false;
